correct range check when reading in tpm commands from a command source
When a message of exactly UTIL_BUF_MAX bytes (currently 8192 bytes) in
the TPM command header is received, then the range check failed to
actually read in the rest of the message. This leads to undefined data
being sent to the tpm.

The tpm device on my machine/kernel rejects this message with an errno
of E2BIG without accessing it apparently.

The call to dump_command() causes the memory to be read, however, which
leads at least to valgrind errors "Invalid read of size ..." but can
also lead to DoS in some situations, probably depending on the details
of heap organization.