Tpm2Command: Check bounds before extracting FlushContext handle.
A malformed FlushContext command that omits the handle requires that we
check the command size before extracting the handle. This commit adds a
unit and integration test for this situation.

This also moves the RM_RC macro from the ResourceManager to the util
header. This macro is useful outside of the ResourceManager and now used
in the Tpm2Command object as well.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>