Tpm2Command: Check bounds before extracting handles.
A malformed command buffer with insufficient size to hold all handles
that the command would normally have requires we check the boundaries
of the command size before extracting handles. This commit adds a
unit test for this situation.

This commit adds bounds checking to the base get / set handle functions
and updates the get / set handles functions (which operate on an array
of handles). We also change the tpm2_command_get_handles prototype such
that the array size is passed by reference. This allows the function to
update this value to let the caller know how many handles were
returned.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>