build: Remove default value for simulator config.
The previous implementation used a default if none was provided by the
caller. If this default (or one provided) doesn't exist then a warning
was printed and the integration tests disabled. This silent disabling of
the tests as the result of what may have been a typo isn't the best
approach.

The new approach is to provide no default. If the caller doesn't provde
a path to a simulator then the integration tests are disabled. If they
do provide a path and it doesn't exist then the configuration is halted
with an error.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>