Thread: Remove pthread_cancel cancelation as default for all threads.
The pattern previously implemented in the Thread class assumed that
calling pthread_cancel followed by an 'unblock' function to free up the
thread would work for the CommandSource, ResourceManager and
ResponseSink. This is however causing deadlocks on older glib versions
(likely related to the GAsyncQueue). The issue seems to be the thread
reaching a cancelation point before it's able to unlock its internal
GAsyncQueue.

This causes the thread trying to tear down the Thread object in question
to block on a subsequent call to thread_wait. This causes the check
target to hang when we wait for the tpm2-abrmd to shutdown. Likely these
hangs were already happening we just weren't catching it.

The changes here include:
Removal of the call to pthread_cancel in the Thread class unblock
function. This requires that the CommandSource class make this call in
a new unblock function. We removed the call to pthread_cancel for the
ResourceManager and ResponseSink, which now rely on their thread
function exiting to terminate. Their unblock functions still use a
control message to unblock them from their GAsyncQueues but the
'process_control_code' function from the util module has been removed.
This was previously used to force a cancelation point in these objects
but it's no longer necessary.

The rest of this patch is cleanup to fixup return values and test code /
cases.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>