Tpm2Response: Wrap all accesses to response handles in range checks.
A malformed response buffer with insufficient size to hold the handle
that it would normally have requires that we check the size of the
response buffer before accessing it. This requires ading a size field to
the Tpm2Response object to enable the check.

Additinoally we add unit tests to be sure these checks are taking place.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>