util: Restructure the handling of return value from 'read' in read_data.
Coverity identified a situation that arises from the read call returning
a negative value that would be propagated to a function as input that
expects unsigned values. This should never happen since read (according
to the man page) only returns -1, 0 or a positive value.

Still, with a minor restructuring of the function we can placate
coverity and make the code better. This requires replacing the switch
statement with an if / elseif / else chain that better suits the
possible return values from the 'read' function.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>