build: Make changes necessary to build in travis docker image.
This requires removing our usage of sudo. We do this by using
PKG_CONFIG_PATH, LD_LIBRARY_PATH and installing the TSS libraries using
the DESTDIR environment variable. Doing so means we don't need sudo to
install stuff, no need to run ldconfig and no need for the weird script
trick we had to do to get the clan compiler on PATH for the root user
when we run sudo to install the TSS.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>