tpm2_command_get_flush_handle: correct error return and invalidate handle
The case of an invalid command code still indicated success to the
caller by returning 0 (which previously was an invalid TPM_HANDLE).
Don't know if TPM_RC_TYPE is the appropriate return value here.

The output parameter handle should be set to a defined value on error to
avoid undefined behaviour if a caller doesn't evaluate error returns
correctly.