travis: Build against TSS master branch & acount for upstream ABI changes.
The commit updates .travis.xml to track the master branch of the TSS.

The 1.x releases of the TSS had the TPMA_SESSION type as a UINT32. The
spec defines this type as a UINT8 however. To preserve backward
compatibility this won't change in the 1.x TSS releases but will be
fixed in a future 2.x release.

Signed-off-by: William Roberts <william.c.roberts@intel.com>
Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>