CommandSource: Set BUF_MAX to PIPE_BUF.
The ResponseSink object currently assumes that all write operations are
atomic. Since we use pipes as the underlying mechanism for sending
responses back to clients we can't assume that writes over PIPE_BUF will
be atomic. This commit is a work around that reduces the maximum buffer
size to PIPE_BUF to guarantee that our write operations will be atomic.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>