build: Add options to the configure script for the systemd preset file.
This allows the caller to specify the default install path for the
preset file, as well as the enable / disable state for the service unit.
By default the install path is `${libdir}/systemd/service-preset`. This
can be overriden using the --with-systemdpresetdir=DIR option. By
default the tpm2-abrmd preset will enable the daemon. It may be disabled
by default using the `--with-systemdpresetdisable` option.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>