tcti-tabrmd: Use local sockets instead of pipes.
The initial implementation of the communication mechanism used to send
command / response buffers between the daemon and the TCTI library was
pipe2(2). This works well but it requires that we manage two fds on both
ends of the communication channel. As we work toward a more generic /
flexible IPC front end getting the implementation down to a single fd
will make things easier to manage.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>