Replace use of socket fds with GSocket.
This is another change intended to decouple the command processing
pipeline from the IPC mechanism. By upleveling the interface from
raw file descriptors to GSocket we'll be able to convert the
CommandSource object to use GSource async I/O instead of polling
the file descriptors directly.

The TCTI library stull has to reach into the GSocket and pull out the
file descriptor to implement the polling / timeout mechanism. Still
using the GSocket in the interfaces allows us to reduce duplication of
code in the util module.

The unit tests were updated to mock the g_socket equivalent of read /
write to preserve test coverage. The test for EINTR on read / write
has been removed as well since GSocket takes care of this for us.

There's also some cleanup in this commit as well since this code
hasn't changed in months and could use some refactoring.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>