CommandSource: Move TPM command buffer read / alloc function to util.
The util module already had most of this functionality. The code in
the CommandSource that allocates the buffer to hold the command buffer
was the only unique code. This code layered quite nicely on top of the
existing read_tpm_buffer function.

This patch also fixes up the tests in the CommandSource unit test.
It also adds unit tests for the new functions to the util unit test.
Finally we've added a few debug / warning messages that were helpful
in debugging this new functionality.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>