CommandSource: Replace use of threads and FDs with GIO objects and a GMainLoop.
Using file descriptors directly makes generalizing the IPC frontend
difficult. The GIO library has objects for doing this (GSource and
callbacks) but they require a thread to drive them. The use of a
pthread for poll/selecting FDs for I/O in the CommandSource was how
we did this manually. But we can't use the same thread to drive the
GIO mechanisms, we need a GMainLoop instead.

So to generalize the I/O stuff we need to replace the threading
mechanism in the CommandSource. Thankfully the Thread interface that
we use to drive threads in the CommandSource (and the other sources
and sinks) works well with GMainLoops too.

This commit combines both of these tasks since they're not really
separable.
- Replace use of FDs to get client I/O through select (2) with GSource
callbacks registered with the G_IO_IN condition.
- Replace pthread internal to the CommandSource with a GMainContext
and Loop object. This thread now drives the I/O loop.

In addition to the changes in the CommandSource, the following objects
were modified to accommodate this work.

Connection:
- Remove remaining use of raw file descriptors.
- The GSocket object is now used as a key for identifying Connection
objects.

ConnectionManager:
- Use new key for Connection object lookup.
- Since the CommandSource now only handles GSockets the
ConnectionManager must be able to find the Connection from the GSocket.

CommandSource unit test:
- Existing tests updated
- New test added to exercise EOF condition in I/O loop.

Signed-off-by: Philip Tricca <flihp@twobit.us>