Command: Replace use of GSocket with GIOStream.
This change touches a number of other objects. The CommandSource and
ResponseSink objects now use GInputStream and GOutputStreams
respectively. The ConnectionManager object is also updated to allow
lookups from the CommandSource using the GInputStream associated with
the Connection.

The util module is also updated to use GInputStream and GOutputStream
objects in place of the GSockets.

Finally all of the unit tests that handle Connection objects or do I/O
using the util functions had to be updated.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>