AccessBroker: Move SAPI context teardown from finalize to dispose function.
This is a corner case where both a gobject and some other resources are
inseperable. Since dispose is where we're supposed to free references to
other gobjects this commit moves the TCTI gobject unref and the freeing
of the SAPI context structure to the dispose function. After this the
finalize function has nothing left to do and so it's removed.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>