Add --flush-all command line option.
This option causes the daemon to flush all transient objects, loaded
and active sessions from the TPM on startup. The purpose of this option
is to put the TPM into as much of a pristine state as we reasonably
can. We only flush tese particular handle type since they're the only
objects managed by the tabrmd.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>