test: Add integration test to create and load sessions over different connections.
The use case behind this feature (and the break in connectionisolation it requires)
is a pipeline or collaborating processes with one creating the session, another using
it. This test explicitly covers the case where the connection that created the
session is still active when a saved session is loaded from another connection.

Access control in this case is outside the scope of the tabrmd. To load the session
from the second connection requires access to the saved context blob (integrity
protected by the TPM). Access to this blob must be controlled by the application
that saved it. Simple filesystem DAC will likely be sufficient for most but use of
SELinux policy is advised.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>