test: Add integration test for continued session LRU algorithm.
This test creates a number of sessions, each over a unique connection to
the daemon, that exceeds (by one) the maximum number that we allow to
continue beyond the life of the connection at once. All sessions are
saved immediately after creation and the associated connection is closed.

When each connection terminates the daemon should:
1) attempt to keep the saved sessions created by the connection around
2) apply an LRU algorithm for flushing the continued sessions when the upper
bound is met

We test this by attempting to load all of the sessions we've saved whos
associated connections are now closed. Since we know the upper bound on the
number of sessions that will survive the LRU and which should have been
flushed (oldest first), we're able to pre-calculate which sessions should
be loadable and which should have been flushed by the daemon.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>