test: Simplify session authroization structure initialization.
Much of the code in our test cases came from the legacy tpmclient test
application. The structure of this code isn't ideal and some of this poor
structure came over with it. Now that I'm getting a better understanding
of the authorization mechanisms and the associated structures from the
TSS2 headers I'm re-writing the structure initialization code. This patch
replaces the convoluted and sometimes unnecessary initialization code
from the `tpmclient` application with designated initializers for the
C structures.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>