test: Remove sapi_fini_from_opts function from context-util module.
Normally wrapping this functionality in a function would make sense but
doing so here causes an unwanted dependency from the context-util module
to the common module. This function is only called from the integration
test driver so it's easier and more clear when we  just call the ~3
functions from the integration test driver directly.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>