SessionList: Use GList instead of GSList & append instead of prepend.
The singly linked list is very limited. We have no reason to not use
the more common doubly linked list.

This commit also changes the insert behavior of the list: instead of
prepending SessionEntries we now append them to the list. This is to
ensure that older SessionEntries are processed by the search functions
first.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>