ResourceManager: Do not flush saved session contexts when connection closes.
If a connection is closed with an associated session(s) having been
saved by the client we no longer flush the session context. Instead
we disassociate the connection from the SessionEntry and transition
this entry into a new state: SESSION_ENTRY_CLIENT_SAVED_CLOSED.

Sessions in this new state will live beyond the lifetime of the
connection that created them and they may be loaded by other client
connections. This requries that the connection attempting to load
the session be in posession of the context blob.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>