ResourceManager: Implement "least recently used" algorithm to limit abandoned sessions.
To support session continuation and limit a potential DoS on abandoned
sessions this commit adds an algorithm that allows only a fixed number
of sessions to be abandoned at a time. Abandoned sessions are those that
were saved by a connection that has terminated but not yet loaded /
claimed by another connection.

We accomplish this by adding a GQueue where we store sessions that have
been abandoned. Sessions for a terminated connection are added to this
queue oldest first. When we add new SessionEntry objects to this queue
we check its length and if we've reached the upper bound we pop and flush
the oldest session.

This algorithm satisfies our desire for "least recently used" semantics
since:
1) When a connection is terminated we process saved sessions from the
oldest to the youngest (stored in this order in the system-wide
SessionList)
2) The GQueue used to hold SessionEntries that have been abanonde has
"last in first out" semantitcs which causes the older SessionEntries to
be removed first.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>