travis: Collect test code coverage metrics & upload to coveralls.
We only enable code coverage stats for GCC since the configure script
will reject other compilers as incompatible with gcov.

Use pip to install the cpp-coveralls package. The COVERALLS_REPO_TOKEN
environment variable is defined in the travis-ci settings pannel for the
project. We use this as a way to determin whether or not to collect
coverage metrics and upload them to coveralls. If you're working on a
private fork that you have connected to travis-ci you can setup a
coveralls instance for your fork and provide this token to your travis
instance to get metrics for your work that hasn't yet gone upstream.

We also do some rearranging here to segregate dependencies from the
tabrmd build. This makes excluding these source trees from the test
coverage metrics easier / cleaner.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>