build: Check that simulator has bound both ports.
This is effectively port assignment collision detection. If the
simulator we start isn't bound to both ports we consider it an error. We
then kill the simulator and try again after an exponential backoff
period. If the simulator is able to bind to the data port but not the
command port it will still run but tests will fail. Killing it off is
important to keep from leaving tpm_server instances around.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>