Remove --fail-on-loaded-trans command line option.
This option was intended as a safe guard against starting the daemon
while some other process had state stored in the TPM. It was an
incomplete implementation however since it fails to address other loaded
objects like sessions. If this functionality is truly needed then it
should be implemented to fail for any loaded TPM state.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>