tcti-tabrmd: Implement config string parsing logic for new init function.
The configuration string format is a series of simple key value pairs
separated by a ',' character. Each key and value string are separated
by a '=' character.

We support two keys:
'bus_name' : The name of the tabrmd service on the bus (a string).
'bus_type' : The type of the dbus instance (a string) limited to
    'session' and 'system'.

Examples:
    "bus_name=com.example.FooBar"
    "bus_type=session"
    "bus_type=system,bus_name=com.example.BarFoo"

This patch includes a pile of new unit tests for this new interface.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>