ResourceManager: Repair session tracking logic when loading abandoned sessions.
When a client loads a previously abandoned session we had previously
neglected to update the structures used to track these sessions. As such
we were getting duplicate SessionEntries in the list of active sessions.
We now detect whether or not the session already exists and update the
internal list of live sessions accordingly.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>