test: Link only against the tabrmd TCTI, all others use dynamic loading.
This cleans out all of the infrastructure that passed the selection of
the TCTI through the TPM20TEST_TCTI_NAME environment variable and the
corresponding device or socket TCTI data through similar env variables.
We also remove all of the TCTI detection and selection logic from the
configure script as well as all of the code that was specific to the
device or socket TCTI.

For all TCTIs that aren't from this repo we use the dynamic loading
stuff that's in the tcti-util module. Configuration for these TCTIs
is passed through the environment like before but using only two
variables:
- TABRMD_TEST_TCTI: this follows the same rules as the --tcti command
line option understood by the daemon
- TABRMD_TEST_TCTI_CONF: that corresponds to the daemon's --tcti-conf
command line option

Removing the bits related to the device and socket TCTI as well as
adding a dependency on libtcti-util.a prompted a move of all LDADD
libraries required for the integration tetsts into a common variable.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>