tcti-dynamic: Don't dlclose tcti libraries till we're done with them.
This resolves a bug in the current implementation. Previously we were
dlclosing the TCTI library after tcti context intitialization. This
caused a segfault on subsquent use of the TCTI context since the library
had been dlclosed / unloaded.

This patch ties the lifetime of the library handle (handle returned by
dlopen) to the lifetime of the TctiDynamic object. There's additional
cleanup in this patch as well. The unit test for the TctiDynamic object
has been split in two adding a unit test for the tcti-util module. These
test cases were previously bundled together.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>