tabrmd: Combine --tcti and --tcti-conf string.
This combineds both into a single configuration string where these two
strings are now combined and separated with a single ':' like
"<name>:<conf>"

Bot the 'name' and 'conf' fields in this string are optional, HOWEVER
if no semicolon is present in the combined conf string it will be
assumed that the whole string is the TCTI name. To provide a 'conf'
string to the default TCTI the first character of the combined
configuration string must be a semicolon.

While breaking the combined configuration string up into its two
component parts, if the name or conf field indicates a default value
then the provided tcti_filename or tcti_conf will not be set. This is to
allow defaults to be set by the caller of the function and only updated
if we're changing them.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>