dist: Add tpmrm to udev rule and allow tss group users to access the TPM2
The udev rule only takes into account the character devices that directly
exposes the TPM2, but no the ones exposed by the TPM zones infrastructure.

Also, it only allows the tss user to access the TPM2 and not the users in
the tss group, so change the permissions so users can also read and write.

Fixes: #334

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>