autotools: Include tpm2-abrmd preset template in distribution tarball
A generated tpm2-abrmd.preset file is included in the distribution tarball, so
there's no way to change the default preset policy by using a configure option.
This commit causes the build to distribute the template file so that
builds from source will generate the final preset file with data from
the configure script. Additionally, the definition of the
systemdpreset_DATA variable is now conditionally defined alongside the
other systemd stuff.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>