ci: Simplify logic / build for coverity.
This commit simplifies the build logic for the coverity build.
All of the VPATH stuff for the coverity build has been replaced with a
simple `./bootstrap && ./configure && make`. The typical build logic in
the `script` section of .travis.yml is now short-circuited when we're
doing a build for coverity since we don't care about the usual build /
test cycle, just the scan.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>