tcti: Remove legacy init function, replace with standard one from spec.
Removing the old init function requires that we use the new one in the
test harness. This new init function takes all configuration parameters
as a string. Using it in the test harness means we no longer pass the
bus type or name through the environment. We now pass the configuration
string through a single environment variable.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>