tcti: Remove tss2_tcti_tabrmd_init_full function.
With the test harness using the standard initialization function now the
need for the tss2_tcti_tabrmd_init_full function is gone. All of the
configuration passed into this function can be passed into the standard
init function in the conf string.

The majority of the churn in this commit is caused by the removal of the
TCTI_TABRMD_DBUS_TYPE enumeration. We had this enumeration in the first
place to keep consumers of the TCTI from having the glib / gio header
pulled into their build (a request from @dguerri). Now that we're using
a configuration string there's no need to hide the GBusType behind our
own enumeration.

Removing this enumeration resulted in a pile of internal changes to use
the GBusType directly. This includes using it in the tabrmd_conf_t
structure. This required fixing up a bunch of unit test code.

Finally this commit also removes the macros that expose default values
for the parameters to the `init_full` function through the installed
header. A NULL conf string will cause the init function to use the right
defaults.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>