tcti: Generalize the key / value parsing from the conf string.
This commit moves the basics of parsing a string of key / value pairs of
the form "key0=value0,key1=value1,..." from the TCTI module to the util
module. This is the part that parses on "=" and ",". To make this
generic we now provide a callback function that the parsing function
will call passing each parsed key / value pair and a reference to some
caller-provided data.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>