test: Combine the two integration test setup scripts into one.
This reduces the duplication between the two scripts and makes the
integration test harness more flexible. This is required for planned
future changes to the daemon to support additional IPC mechanisms.

This required adding an additional option to the test script:
'--tabrmd-tcti'. This option allows the caller (or the make file) to
specifiy which TCTI the daemon will use to communicate with the TPM.
This can be either 'mssim' or 'device'. We've also fixed up the
Makefile.am to construct the proper command line & environment for
the consolidated test script based on the options passed to the
configure script.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>