test: Fix inaccurate number of handles for loaded transient objects.
No clue why I was adding '1' to the count here, probably thought that I
had to account for a 0 based index. This wasn't correct however. The end
result was that the 'get_transiet_handles' function in the
get-capability-handles-transient.int test was returning a count
indicating that the array holding the handles had one more handle than
was actually there. This doesn't fail on most systems since the array
was likely 0 initialized and the test just compared 0 to 0 and didn't
report the failure. It will fail however if the compiler doesn't
initialize the handle array to zero.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>