INSTALL.md: Add section for "sanity checking" the daemon.
There have been a lot of problems reported recently as people have been
trying to debug connection issues between clients (generally one of the
tpm2-tools executables) and the daemon. These usually boil down to a
configuration error when building the tools but the daemon is often
blamed since people have a hard time verifying that it's functioning
properly. Most of this is due to the relatively opaque nature of dbus
and the fact that most people don't know how to query dbus from the
command line.

This new section added to the tail end of the INSTALL.md document
attempts to fill this void with some data w/r/t invoking the `dbus-send`
utility. Specifically this patch adds instructions for executing a
method that causes the DBus daemon to list all of the names that daemons
have claimed on the bus. This allows people to verify that the daemon
has clamed the right name. Additionally we provide another `dbus-send`
command to query the `Introspect` interface to get the tabrmd to dump
XML describing the functions / methods that it supports.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>