build: Replace --with-simulatorbin option with --enable-integration.
The --with-simulatorbin option was a bit awkward as it required passing
the absolute path to the simulator executable. Just having the configure
script check the callers PATH to make sure it's there is a lot cleaner.
This implies that if `--enable-integration` is passed to configure the
caller must have the `tpm_server` executable on their PATH.

Additionally the configure script must export the PATH as it was when the
configure script was executed. The Makefile must then propagate the PATH
to the test script as its executed.

The INSTALL.md file is updated to describe all of this as well as to
document the interaction between this new option with the
`--enable-test-hwtpm` option.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>