ResourceManager: Add new session to list of loaded sessions when mapping.
The logic in the 'create_context_mapping_session' function was causing
us to run out of space for loaded sessions. This in turn caused the
extended session logic to break down since those tests load and unload
a lot of sessions.

The problem was the logic that detected which of the internal data
structures held a SessionEntry after a new session was created. The
ResourceManager wasn't adding it to the list of loaded sessions and so
the ResourceManager wouldn't unload them. This caused us to run out of
space for loaded sessions very quickly.

This commit also adds a comment block above the SessionEntry state
enumeration to make it clear what the states mean and the constraints
that apply when a session is in a given state.

Finally this commit also moves the test/integration/auth-session-max.int
test back to the group of expected failures (XFAIL). The bug fixed above
is what caused this test to pass and now that it's fixed this test to
fail again.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>