build: Treat warnings as errors when checking compiler for flag support.
The command executed by the AX_CHECK_*_FLAG macro will cause clang to
report a warning when a flag isn't supported. But if we pass this flag
to the build the compilation will fail. So we have to force the warning
to be treated as an error when AX_CHECK_*_FLAG is executed by adding
-Werror to CFLAGS for the check.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>