Revert "Add scan-build Static Code Analysis to Travis"
This reverts commit f05f2a850efc88b99e87da31a2316053fe0fe45f. The issue
with this commit is that, while it enables scan-build, it overwrites the
contents of CFLAGS and causes the build to fail (all -I directives to
the compiler have been lost and so it can't find the TSS@ headers). Odly
enough, because scan-build didn't find any errors (the build failed so
no errors to find) it still reports success. This isn't what we want.

The right solution here will preserve CFLAGS. A simple change like this
however causes the build to pass complete, scan-build to run, but it
then finds a pile of issues and so the build fails. For the short-term
we're reverting this commit and will resubmit it once the issues
identified by scan-build are resolved.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>