test: Convert create-keys into a test for max-transient option.
This test used to just create a primary key and then 5 keys under it.
Now it attempts to create and load the default maximum of transient
objects plus one more. The final attempt to load a key should fail. If
it doesn't, or if it fails in any other way then the test will fail.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>