ResourceManager: Delete link from abandoned queue after it's taken over.
Searching through a GQueue isn't a particularly clean process. Using the
'find' function is easy enough but once we find the session we're
looking for we must remember to remove / delete the associated link.
Otherwise the link will remain in the queue and we end up with a few
problems as a result.

Firstly the upper bound on the number of abandoned sessions will get
screwed up since we won't actually be removing ones that are claimed by
other connections. Secondly we'll could up with a double g_object_unref
when the ResourceManager is torn down. Removing the link after we look
it up prevents all of this.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>